BIS-51 / SP-139: Bugfix: Unnecessary calls to RelationshipService.assignSampleToContainer() and RelationshipService.removeSampleFromContainer()
SVN: 25983

+++ openbis/source/java/ch/systemsx/cisd/openbis/generic/server/business/bo/AbstractSampleBusinessObject.java
@@ -210,6 +210,16 @@ abstract class AbstractSampleBusinessObject extends AbstractSampleIdentifierBusi
         final SamplePE containerPE =
                 tryGetValidNotContainedSample(containerIdentifier, sampleIdentifier, defaultSpace);
 
+        if (samplePE.getContainer() == null && containerPE == null)
+        {
+            return;
+        }
+
+        if (samplePE.getContainer() != null && samplePE.getContainer().equals(containerPE))
+        {
+            return;
+        }
+
         if (containerPE == null)
         {
             relationshipService.removeSampleFromContainer(session, IdentifierHelper
